Remove automatic table of contents
